+++ src/traits/routing_trait.rs
@@ -16,36 +16,17 @@
     See the Licences for the specific language governing permissions and limitations relating to
     use of the MaidSafe Software.                                                                 */
 
-pub fn array_as_vector(arr: &[u8]) -> Vec<u8> {
-  let mut vector = Vec::new();
-  for i in arr.iter() {
-    vector.push(*i);
-  }
-  vector
-}
+/// temporary code to test passing a trait to routing to query and possible decode types or
+/// at least soem info routing needs which is access to these functions on data types
+/// These traits will be defined in routing and require to be avauilable for any type
+/// passed to routing, refresh / account transfer is optional
+/// The name will let routing know its an NaeManager and the owner will allow routing to hash
+/// the requsters id with this name (by hashing the requesters id) for put and post messages
+use common::NameType;
 
-pub fn vector_as_u8_32_array(vector: Vec<u8>) -> [u8;32] {
-  let mut arr = [0u8;32];
-  for i in (0..32) {
-    arr[i] = vector[i];
-  }
-  arr
-}
-
-pub fn vector_as_u8_64_array(vector: Vec<u8>) -> [u8;64] {
-  let mut arr = [0u8;64];
-  for i in (0..64) {
-    arr[i] = vector[i];
-  }
-  arr
-}
-
-#[allow(dead_code)]
-pub fn compare_arr_u8_64(lhs: &[u8; 64], rhs: &[u8; 64])->bool {
-  for i in 0..64 {
-    if lhs[i] != rhs[i] {
-      return false;
-    }
-  }
-  return true;
+pub trait RoutingTrait {
+	fn get_name(&self)->&NameType;
+	fn get_owner(&self)->&Vec<u8>;
+	fn refresh(&self)->bool { false } // is this an account transfer type
+	fn merge(&self)->bool { false } // how do we merge these
 }
